DYN-4240-Step5-SearchResults (#12250)
* DYN-4240-Step5-SearchResults

I added the needed resources for the title and content of the Step 5 (Search Results) and Step 6 (Package Dependencies).
Also I updated the RichTextBox behavior for showing the bullet points in the middle of the text.
Also I had to modify the Style for the ViewDetails button due that we need a grid in order to highlight the button with the orange rectangle.
Finally I updated the json file for showing the Step 5 and Step 6 and implemented the methods for the corresponding automation that should be executed for the Step 5. When the ViewDetails button is clicked or when the Popup next button is clicked we should be moved to the next Step also the View Details button should be highlighted.

* DYN-4240-Step5-SearchResults Code Review

Some comments in the RichTextBox were added and also a variable was removed due that was not used.
Also in the GuidesManager some comments were updated.

* DYN-4240-Step5-SearchResults Conflicts

After merging the changes from the master branch, the Packages guide was not reachable until Step 6, then I noticed that a sequence was repeated in the json and there was no prevalidation then I updated the json with the right Sequences.