DYN-4241-Step6-PackageDependencies
I added the Step 6 to the packages guide, the automation needed for opening the PackageDetaiils and adding support for showing the PackageDetails User Control over the dark overlay.
The highlight and cutoff functionality was moved from the Guide class to the Step class, so it will be executed on Step.Show(), Step.Hide() or when exiting the Guide.
Also I modified resources for the Package Details popup and remove the resource for the "Autodesk Sample Package" due that it shouldn't be localized.
I added some code to subscribe the "Search for a Package" menuitem Click event to the NextStep method.
I disable the Back button in the Step1 due that was closing all the Steps and leaving the dark overlay.
Finally I added some methods for closing the PackageSearch window