DYN-4276-Step8-NavigateToPackages
I've added the resources for the text and title of the Steps 8, Step 9 and Step 10 (survey) and the json structure.
I've added the functionality for putting the background, the cutoff area and the highlight rectangle over the LibraryView when showing Step8.
Changed the Access Modifiers for CurrentStep and NextStep due that we need to use both in the LibraryViewExtensionMSWebBrowser.
Also I added some methods for subscribe events in the Step8 Popup Next button, for subscribe the click event in the Package and for showing/hiding the dark overlay.