DYN-4276-Step9-NavigateToPackages
I updated a resource string for the Step2 - FindAPackage (reported by Aabi).
I merged the latest changes in master related to the LibraryView UX
Also I modified the way we are showing the cutoff section in the Library due that with the latest changes to the LibraryView was appearing incorrectly.