DYN-4276-Step9-NavigateToPackages Code Review
The method ExecuteJSFunction was moved from Guide.cs to ResourceUtilities.cs.
The harcoded text "NextButton" was moved to the json file so we can create unit test easly and also for subscribing the Button.Click event is using the ManageEventHandler method.
Some properties that were public due that were used in LibraryViewController were changed back to internal.
The handlers events used for moving to the next step or to the previous step were updated so now we won't be passing the Step.Sequence as parameter. This will allow to call the NextStep event without passing arguments so it can be called from the LibraryViewController.cs