DYN-4276 Step8 Navigate to Packages (#12312) (#12366)
* DYN-4276-Step8-NavigateToPackages

I've added the resources for the text and title of the Steps 8, Step 9 and Step 10 (survey) and the json structure.
I've added the functionality for putting the background, the cutoff area and the highlight rectangle over the LibraryView when showing Step8.
Changed the Access Modifiers for CurrentStep and NextStep due that we need to use both in the LibraryViewExtensionMSWebBrowser.
Also I added some methods for subscribe events in the Step8 Popup Next button, for subscribe the click event in the Package and for showing/hiding the dark overlay.

* DYN-4276-Step8-NavigateToPackages

I refactored some functions so the function that use reflection can be reused.
I added several comments to the new methods created.

* DYN-4276-Step9-NavigateToPackages

I updated a resource string for the Step2 - FindAPackage (reported by Aabi).
I merged the latest changes in master related to the LibraryView UX
Also I modified the way we are showing the cutoff section in the Library due that with the latest changes to the LibraryView was appearing incorrectly.

* DYN-4276-Step8-NavigateToPackages Code Review

Instead of calling each javascript method in GuidesValidationMethods.cs I added a new section in the json file for executing javascript functions.
I modified some comments and also some fixes in the library.html to make the code better.

* DYN-4276-Step9-NavigateToPackages

Fixing bug related to allowing the user to interact through the overlay, when the overlay is shown then no user interaction should be allowed.

* DYN-4276-Step9-NavigateToPackages Code Review

The method ExecuteJSFunction was moved from Guide.cs to ResourceUtilities.cs.
The harcoded text "NextButton" was moved to the json file so we can create unit test easly and also for subscribing the Button.Click event is using the ManageEventHandler method.
Some properties that were public due that were used in LibraryViewController were changed back to internal.
The handlers events used for moving to the next step or to the previous step were updated so now we won't be passing the Step.Sequence as parameter. This will allow to call the NextStep event without passing arguments so it can be called from the LibraryViewController.cs

* DYN-4276-Step9-NavigateToPackages Code Review

I added some code related to the Step 10 (survey), a resource was updated and a line in GuideManage.cs
Also I moved some hardcoded strings from the GuidesValidationMethods class to the json file so we can create unit test that validate the js functions existing in the html file.