Step8 Wrong Popup Location
I added some methods in the library.html so we can scroll down in the library and also for getting all the information related to a specific item.
I modified the code for updating a popup location so if the Popup is located over the Library we will be recalculating the popup location if the target library item location is changed.
Also I added a new event that will allow to update the current Popup location
Finally I modified the ExecuteJSFunction method so we will get the value returned by the JS function.