DYN-4422-Step7-GuideCrash
The crash was happening when coming from  Step7 to Step6 and the sample package was not installed, it was happening because Step6 is expecting the PackageSearch window opened but when going from Step7 to Step6 it was closed, then Step6 was disappearing (Popup.PlacementTarget doesn't exist).
Then this fix is preventing to close the PackageSearch window when coming from Step7 to Step6 so the Step6 will be appearing right.
Also I included the changes related to the new package name uploaded "Dynamo Samples" and one of the contents "SampleLibraryUI" .