DYN-4276-Step8-BugFix2
When the sample package is installed for the first time the Step8 Popup was in the wrong location then I did a change to that then the package is installed the Popup location and interactions are updated.
For implementing this I added a new event called OnUpdateLibraryInteractions() and also added a method that will be subscribing the package click event and putting the Highlight over the Library.
The IsAnyGuideActive property needs to be public due that will be using in a different assembly.