DYN-4408-UILayout-Issue
I reduced the HorizontalPopupOffset value in the json file due that now the HorizontalOffset is autocalculated automatically based in the MenuItem.Popup width (just then there is an UIAutomation that is executed for opening the MenuItem).
For calculating the HorizontalOffset first we need to get the MenuItem.Popup border so we can get the Width.