DYN-4252-Shadow-Under-Pointer (#12503)
* DYN-4252-Shadow-Under-Pointer

I had to replace the Polygon (tree points connected by 3 lines) by a Path (tree points connected by 2 lines) due that the shadow was being applied to all sides of the Polygon(triangle), then now due that is a Path with 2 lines the shadow will be applied just in two lines.
I had to add a Converter so a PointCollection can be converted to a Path.Data.

* DYN-4252-Shadow-Under-Pointer Code Review

Fixed one comment