DYN-4135-UnpinFromNode
I added the new option "unpin from node" in the Note Context Menu and also added the needed resources.
Also I added a new method CanUnpinFromNode so it will manage when the option will be enabled and disabled in the context menu.
Finally I did a fix in NoteView.xaml.cs due that every time that the right click was pressed over the note a new DoubleInput (number) node was added to the selection so it was making the CanUnpinFromNode() method to be non working in some cases.