DYN-4412-Filters-PackageManager
I did a fix in PackageManagerSearchViewModel.cs, so the filter after getting the result will be doing an intersection instead of a union.
Also I added a unit test that will adding 9 packages in total, 3 are from FormIt, 4 are from Advance Steel, and 2 are from both Formit and Advance Steel, then after applying the filters Formit and Advanced steel we will get only the two of them that belong to those two categories.