DYN-3744-ByCoordinatesNode-DocBrowser (#12608)
There was an issue in the message warning that appear over the node when connecting a CodeBlock node with a string inside and a Point.ByCoordinates node due that the input was expecting a number not a string, then in the warning message it was missing the "Learn more" link and it was due that the ResourceString was missing the webpage name. Then I just added the webpage name at the end of the string and now the warning message is showing the "Learn more" link.