DYN-4552-Pick-GroupStyles
With the next changes are adding a Menu entry in the Group ContextMenu, now it will be showing all the GroupStyles added in the Preferences panel, so if a style is added or removed, the ContextMenu GroupStyle will be updated.
Basically I created one base class named GroupStyle base and two child classes GroupStyleItemEntry and GroupStyleSeparator, those classes will be used to select the right Style depending if is a MenuItem or a Separator.
Also I added several resources to avoid using hard-coded strings.
I modified the AnnotationView for adding an entry in the ContextMenu that will be showing all the GroupStyles, also I added a Style for the MenuItem so it can show the scrolbar when there are several MenuItems.
I added the logic for the GroupStyles context menu, this are the rules:
- If a Custom GroupStyle is selected and the user clicks it then the Default Style will be selected.
- If no GroupStyle is selected and the user clicks a specific GroupStyle then it will be selected and the style will be applied to the Group.
Finally I added two unit tests, one is validating that the Default Styles exists and the other one will be adding a style and checking that is shown in the Group ContextMenu.