DYN-4552-Pick-GroupStyles CodeReview3
Moved the functionality of creating the Default Styles from PreferencesViewModel to the PreferenceSettings class. the call to AddDefaultStyles() method cannot be done in the PreferenceSettings constructor due that the constructor is executed first and then the Serialization process, so the default styles were added several times after reopening Dynamo, that's why the call to  AddDefaultStyles() need to be done in DynamoModel.cs after the serialization process was already done.
Finally I moved all the resources from DynamoCoreWpf project to DynamoCore project.