GroupStyle ContextMenu crashing
A crash was reported when opening the Group ContextMenu using the three dots, when opening the ContextMenu by clicking right over the Group was not crashing.
The problem was due that the Styles were at the level of <UserControl.Resources> then the method call FrameworkElement.FindResource() was failing (Resource not found) then I moved the Styles to the scope of <MenuItem.Resources> so they are always found (I've tested both cases of showing the ContextMenu).