DYN-4685-Fix-PinButton-AddStyle
I modified the style for the PinIcon for following the HIG standards also I modified the style for the "Add Style" button due that was not consistent with other buttons in the Preferences window. For the "Add Style" button I've added a Storyboard for modifying the Border.Background when the button is pressed.