DYN-4743-GroupStyles-SelectionState
Now every time that a GroupStyle box is selected a blue border of 2 px will be shown, the default ones cannot be selected just the custom ones, when the mouse is over a GroupStyle box a gray border of 2px will be shown (except for the default ones).
Also if a custom GroupStyle is selected and the user clicks over other parts of the Group Style section, the current selected one will be deselected.
Finally I did some re-organization in the layout so all the controls are vertically aligned (due that after showing the selected/mouseover border it was moving the groupsstyle boxes in a weird way).