DYN-4764-GuidedTour-Step3-Nodes
I've added several properties in the CutOffArea class that allow to move the cutoff section over the DynamoWindow also a property that will hold the Node ID useful when searching nodes in the Workspace.
The Step3-Nodes was added to the json file so it will show the Popup with the desired information.
I re-organized some methods already existing in the Guide class and were moved to a new class named GuideUtilities (so in all the places in which those methods were used were updated).
A new method was added in GuidesValidationMethods, this method will put the highlight rectangle over the port..
In the UIAutomation class a new property was added, this will be used for the cases when a function is executed we can pass parameter from the json file.