DYN-4766-Step4-ConnectTheNodes
Removing the file "nodePackage - Copy.html" due that was a copy inserted by mistake
Adding resources for the Step5 (in this way we can test the navigation from Step5 to Step4).
Modified the json for adding Step4 and Step5 and fixing the location of the user added node.
Updated the dyn file due that now the ByOrigin node won't be present at the beginning of the tour.
I created several methods in GuideValidationMethods for Removing the user created connection and also some function for getting nodes.
Also updated several methods in GuideValidationMethods for using the nodemodel "lastCreatedNode", added in previous steps.
In GuideUtilities I added a new method for getting all the nodes of a specific type.
The Point2D NodePosition was changed to public again due that it was not serializing the info, so the node was added always in the 0,0 position.