DYN-4766-Step4-ConnectTheNodes Code Review
Removed the emptyImage.png due that was deleted in a previous pull request.
Instead of using ToString().Replace I'm passing a parameter to the ToString() method.
The hardcoded value "CoordinateSystem.ByOrigin", now is in the json file and passed as parameter.
Added some lines for validating the number of parameters.