DYN-4759-FileTrustUI-Popup
I modified several method is HomeWorkspaceModel, DynamoModel, DynamoModelCommands and RecordableCommands due that I added the parameter forceAutomaticWithoutRun so when we open a file with RunMode = Automatic it won't be executed automatically.
In the RunSettings class I've added some properties so we can enable/disable the ComboBox RunType and the tooltip also modified the xaml to make the bindings and show the right style when the RunButton is disabled.
I've added the resources needed for the UI and also the style for the buttons and the checkbox.
By requirement (described in the Jira task) I've added an entry in the Debug menu that will shown the Warning popup when there is no file opened.
Finally I've added WPF Popup needed for showing the File Trust Warning Popup when a new dyn file is opened just if the file path doesn't exists in the trusted locations.