DYN-4759-Fixing-BreakingChnges
Removing the forceAutomaticWithoutRun parameter added in OpenFileCommand and OpenFileFromJsonCommand in order to not produce breaking changes.
I've added a new command called OpenFileAutomaticWithoutRun that will be used when opening a graph with RunMode = Automatic but won't be run.

Removing the forceAutomaticWithoutRun parameter added in the next public functions due that is causing breaking changes:
-  public virtual void ResetEngine(bool markNodesAsDirty = false)
- public void OpenFileFromJson(string fileContents, bool forceManualExecutionMode = false)
- public void OpenFileFromPath(string filePath, bool forceManualExecutionMode = false)

Also removed the forceAutomaticWithoutRun parameter from the next private methods.
- private bool OpenJsonFileFromPath
- private bool OpenXmlFileFromPath
- private void SetPeriodicEvaluation