DYN-4759-RunAutomatic-Bug
A bug was reported during testing due that when opening a dyn file with RunMode=Automatic the graph was executed even when showing the file trust warning.
The property ForceAutomaticWithoutRun was converted to be static due that is used in several places and we need to use it before and after executing the OpenFileCommand, due that the call for opening the file is async and the RunSettings instance is created two times one when we open a dyn file and the other one when dynamo is opened then this property needs to be static.