DYN-4999-YesButton-RunGraph
With this changes the Run button and RunTypes ComboBox will be enabled after pressing the Yes button in the FileTrust Warning (no matter the status of the checkbox), by consequence of this change the  ComboBox tooltip will be disabled(due that it's Visibility is binded to the RunTypesEnabled property).
Also when the Yes button is pressed (no matter the RunTypes ComboBox) a run will be executed.

Also for dyn files with RunMode = Periodic the RunTypes ComboBox will be disabled and will be enabled once we press Yes button and also a run will be executed.