DYN-4973-Popups-NotAttached
I implemented the functionality of  hiding/showing the FileTrustUI and GuidedTours popups when the Dynamo window is inactive (in background).
I've tested the functionality for the User Interface tour and Packages tour and for the exitwindow that appear in the Packages tour.
For the other tours the RealTimeWindow (that appear when exiting the Guide) wont' have this functionality implemented.