DYN-4833-Group-NodesAlignment
Before this fix all the  alignment calculations were considering all the nodes selected then also was considering the Group (AnnotationModel), so when aligning to Left, Right, X-Distribute, Y-Distribute the Group was being moved some pixels so the right, left or over X-axis and Y-axis.
My fix was not considering the Group in the selected items so if we are trying to align all the nodes inside the Group, the group location won't change (just if is necessary in some cases like Left or Right alignment).