DYN-5106-WebView2-DocumentationBrowser-Fix1 (#13256)
When displaying the documentation for NodeAutocomplete the GIFs were not showing correctly because the base64 content was filtered by HttpUtility.UrlDecode method then I removed this method call and also added code for joining the two sections of the URL, for this I'm using a call to  string.Join. After this fix the images for nodes and for node autocomplete are shown correctly.