YN-5118-WebView2-Library Fixing interaction with Packages guide
Some functionalities when interacting with the packages guide were broken like:
- highlight installed package (with glow animation).
- scrolldown automatically.
- show/hide dark overlay over Library
- show/hide Library when is not needed.
- Locate the Guide popup in the right place vertically next to Library

Basically the changes are to now use the WebView2.ExecuteScriptAsync() method and  passing the parameters correctly, additionally some methods were converted so async for supporting ExecuteScriptAsync() method.
Also I noticed that an extra argument for the js method findPackageDiv was added but not all the methods were updated so some functionalities were invalidated due to that.
Finally for scrolling down automatically the Library I had to find a specific section of the Library and then scrolling down, because otherwise was not working.