Dyn 5317 check webview2 installed (#13464)
* DYN-5317-Check-WebView2-Installed

I added a validation that before starting the DynamoSandbox app we will check if the WebView2 Runtime is installed or not, if is not installed we will show a MessageBox with the  link of the WebView2 installer and then after pressing the OK button close Dynamo.
Also in the DynamoMessageBox I added a RichTextBox that will be visible under very specific circumstances, this RichTextBox allows to have a blue hyperlink as a part of the text so the user can click it and open the webview2 installer webpage.

* DYN-5317-Check-WebView2-Installed

removing extra comment