DYN-5286-MLNodeAutocomplete-Analytics CodeReview1
Removed the Actions added in IAnalyticsClient.cs so we can re-use the existing ones (so the method calls were also updated).
Added Analytics track events in the Preferences panel when the NodeAutocomplete is se to ML Recomentations or Object Type.
The ExportSettings option and ImportSettings option were renamed to Export and Import (so the method calls were also updated).