DYN-5497-Markdown-NotLoading CodeReview2
The Name property was named to packageName
A new test was added that is validating that the virtual folder will be created in the packages/package/doc folder so when the documentation for a node is loaded the images are loaded successfully.
Also the code for setting the VirtualFolderPath property was moved before WebView2.EnsureCoreWebView2Async() call because otherwise we were not reaching the code.