DYN-5497-Markdown-NotLoading MD for Orchid nodes
There are some Orchid nodes in the Math category that are receiving a number of parameters variable and seems that node documentation (md files) is not supported for this kind of nodes (DSVarArgFunction), then a fix was applied for loading the package info and md files for this kind of nodes.
Also I fixed a test that was replacing slashes but in some cases the slashes were already inverted.