DYN-4738-Custom-ColorPicker Fix
Due that I was using the code " xctk:ColorPicker" in several places inside the Style/Templates defined in DynamoModern.xaml, the xaml parser was crashing when executing a test located inside in DynamoCoreWpfTest (I think was because this project doesn't have the Extended.Wpf.Toolkit package installed).

Then the fix was moving the Styles defined in DynamoModern.xaml to CustomColorPicker.xaml (in this project the Extended.Wpf.Toolkit package is installed).
By consequence also the two converters added in DynamoConverters.xaml were also moved to CustomColorPicker.xaml