DYN-5742-Add-Workspace
Removing previous fix because it was incorrect, the problem is that when executing DynamoSandbox from a path that contain empty spaces the URI.AbsolutePath is converting them to "%20" so the dyn file is not found.
Then with this fix I'm removing the "%20" from the string path and is found correctly when inserting the graph in the workspace.