DYN-5732-PythonNodes-HelpDoc
Fixing the next bugs for the PythonScript and PythonScriptFromString:
The preview graph images in DocumentationBrowser are not updated using the latest modern style.
When trying to insert the Documentation graph into the workspace is sending an error due that the dyn file doesn't exist.

Then I did the next fixes:
- The missing dyn files were added to Dynamo\doc\distrib\NodeHelpFiles folder.
- The dyn files were updated in Dynamo and got screenshots for replacing the jpg images in Dynamo\doc\distrib\NodeHelpFiles.