DYN-5844-ColorPicker-MarginIssues
We have a Canvas inside a Grid but the Canvas doesn't support auto-size like other WPF controls then the only way to fix the problem of the CustomColorPicker appearing wrong under a 200% scale is to use a fixed height.
In this way this fix will make the NormalColors window with the same height than the CustomColors window.