Fixing Regressions
Adding the Input and Output nodes are breaking the tests (due that we have two input and two output nodes), then removing the indexing of this nodes.

Other tests were failing due that writer?.Dispose() disposed the object but later (when opening the dyn file) is trying to index again the nodes and the writer instance was already disposed.