DYN-6049 Fixing Github Actions Security
After a external researcher ran a vulnerability scanner over Dynamo github actions found vulnerabilities with the variable env.event.issue.title due that was used as a parameter for a powershell script (after was parsed), then the suggested fix (https://securitylab.github.com/research/github-actions-untrusted-input/) was to use a local environment variable and then instead use the environment variable as parameter for the script.
So I've applied this fix in several github actions that had the same security bug.