PackageManager Lucene Crash
After opening the Package Manager and closing it was crashing due that the LuceneSearchUtility.Writer was null (the second time we don't reach the PackageManagerSearchViewModel constructor so the InitializeLuceneConfig() was not called).
Then for this fix if the PackageManagerSearchViewModel was already created then we have to call again the InitializeLuceneConfig().