DYN-6061 Search Missing Nodes
Due that we were using a StandardAnalyzer for Lucene searching  we were not able to find some nodes in the results like "+", "*", "And". So I had to implement a Custom Analyzer so we will be using a specific Tokenizer that supports those special characters in the search term.
Also I've added a Unit Test that validates if a set of specific nodes are found or not.