Fix Node Autocomplete Regressions
Due that we are adding ZeroTouchNodes during test mode around 9 tests are failing, seems that there are several nodes duplicates so each test needs to be analyzed separately.
Then in this fix I'm reverting the code for adding ZeroTouchNodes and also removing a test that will fail due to this change.