DYN-6073 Civil3D Package Tour Crashing
This change is disabling the close button in the PackageManagerSearch window when running the Packages tour, when passing to the next Step the button is enabled again (unless the next step also requires to disable the button). I've added a new icon image that will be shown when the close button is disabled.
In this way we will be preventing the crash when the user try to close the Packages tour by closing the PackageManagerSearch window