DYN-6073 civil3 d packages tour crashing (#14338)
* DYN-6073 Civil3D Packages Guide Crashing

It was crashing due that the user tried to close the Step 4 clicking the PackageSearch window and the Popup is not closed but seems that is already disposed.

* DYN-6073 Civil3D Package Tour Crashing

This change is disabling the close button in the PackageManagerSearch window when running the Packages tour, when passing to the next Step the button is enabled again (unless the next step also requires to disable the button). I've added a new icon image that will be shown when the close button is disabled.
In this way we will be preventing the crash when the user try to close the Packages tour by closing the PackageManagerSearch window

* DYN-6073 Civil3D Package Tour Crashing

This fix will solve the problem of the packages guide crashing when clicking the Library (package installed) for passing from Step8 to Step9.